refactor: Reorganize project structure for better maintainability
This major reorganization improves the Arduino project structure while
maintaining full compilation compatibility:

## Structure Changes
- src/: Core application code organized by function
  - network/: WiFi management and weather API client
  - display/: UI components and graphics handling
  - security/: SSL certificate management
- assets/: Non-code resources (fonts, images, certificates)
- tools/: Development utilities (image converter, cert collector)
- docs/: Project documentation

## Technical Improvements
- Fixed library compatibility issues (DisplayLocations.h, InkplateLibrary)
- Updated all #include paths to reflect new structure
- Maintained Arduino IDE compatibility requirements
- Preserved git history through intelligent file renames

## Benefits
- Clear separation of concerns by component type
- Easier navigation and code discovery
- Scalable structure for future additions
- Professional project organization

Compilation verified: 57% program storage, 17% dynamic memory (unchanged)
