+++ src/display/KittyPics.cpp
