+++ src/network/CurrentConditions.h
